services/sessions/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyjwt==2.10.1
